+++ tests/event/listener_base.php
@@ -42,7 +42,7 @@ class listener_base extends \phpbb_test_case
 	/** @var string */
 	protected $bbvideo_height;
 
-	public function setUp()
+	protected function setUp(): void
 	{
 		parent::setUp();